Rename authority provider type from STORAGE to NATIVE
