+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-common/src/main/java/org/apache/shardingsphere/authority/algorithm/natived/loader/StoragePrivilegeLoadEngine.java
@@ -15,9 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.authority.algorithm.storage.loader;
+package org.apache.shardingsphere.authority.algorithm.natived.loader;
 
-import org.apache.shardingsphere.authority.algorithm.storage.loader.impl.StoragePrivilegeBuilder;
+import org.apache.shardingsphere.authority.algorithm.natived.loader.impl.StoragePrivilegeBuilder;
 import org.apache.shardingsphere.authority.model.ShardingSpherePrivileges;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
 import org.apache.shardingsphere.infra.metadata.user.ShardingSphereUser;
