+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-common/src/main/java/org/apache/shardingsphere/authority/algorithm/natived/loader/impl/dialect/OraclePrivilegeLoader.java
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.authority.algorithm.storage.loader.impl.dialect;
+package org.apache.shardingsphere.authority.algorithm.natived.loader.impl.dialect;
 
-import org.apache.shardingsphere.authority.model.database.SchemaPrivileges;
-import org.apache.shardingsphere.authority.model.database.TablePrivileges;
-import org.apache.shardingsphere.authority.algorithm.storage.loader.StoragePrivilegeLoader;
+import org.apache.shardingsphere.authority.algorithm.natived.loader.StoragePrivilegeLoader;
 import org.apache.shardingsphere.authority.model.PrivilegeType;
 import org.apache.shardingsphere.authority.model.ShardingSpherePrivileges;
+import org.apache.shardingsphere.authority.model.database.SchemaPrivileges;
+import org.apache.shardingsphere.authority.model.database.TablePrivileges;
 import org.apache.shardingsphere.infra.metadata.user.Grantee;
 import org.apache.shardingsphere.infra.metadata.user.ShardingSphereUser;
 
@@ -34,7 +34,6 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -42,38 +41,37 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 /**
- * PostgreSQL privilege loader.
+ * Oracle privilege loader.
  */
-public final class PostgreSQLPrivilegeLoader implements StoragePrivilegeLoader {
-    
-    private static final String ROLES_SQL = "select * from pg_roles WHERE rolname IN (%s)";
-    
-    private static final String TABLE_PRIVILEGE_SQL = 
-            "SELECT grantor, grantee, table_catalog, table_name, privilege_type, is_grantable from information_schema.table_privileges WHERE grantee IN (%s)";
-    
+public final class OraclePrivilegeLoader implements StoragePrivilegeLoader {
+
+    private static final String SYS_PRIVILEGE_SQL = "SELECT GRANTEE, PRIVILEGE, ADMIN_OPTION, INHERITED FROM DBA_SYS_PRIVS WHERE GRANTEE IN (%s)";
+
+    private static final String TABLE_PRIVILEGE_SQL = "SELECT GRANTEE, TABLE_SCHEMA, TABLE_NAME, PRIVILEGE, GRANTABLE, INHERITED FROM ALL_TAB_PRIVS WHERE GRANTEE IN (%s)";
+
     @Override
     public Map<ShardingSphereUser, ShardingSpherePrivileges> load(final Collection<ShardingSphereUser> users, final DataSource dataSource) throws SQLException {
         Map<ShardingSphereUser, ShardingSpherePrivileges> result = new LinkedHashMap<>();
         users.forEach(user -> result.put(user, new ShardingSpherePrivileges()));
+        fillSysPrivileges(result, dataSource, users);
         fillTablePrivileges(result, dataSource, users);
-        fillRolePrivileges(result, dataSource, users);
         return result;
     }
-    
-    private void fillTablePrivileges(final Map<ShardingSphereUser, ShardingSpherePrivileges> userPrivilegeMap, 
-                                     final DataSource dataSource, final Collection<ShardingSphereUser> users) throws SQLException {
+
+    private void fillTablePrivileges(final Map<ShardingSphereUser, ShardingSpherePrivileges> userPrivilegeMap, final DataSource dataSource,
+                                     final Collection<ShardingSphereUser> users) throws SQLException {
         Map<ShardingSphereUser, Map<String, Map<String, List<PrivilegeType>>>> privilegeCache = new HashMap<>();
         try (Connection connection = dataSource.getConnection()) {
             Statement statement = connection.createStatement();
             try (ResultSet resultSet = statement.executeQuery(getTablePrivilegesSQL(users))) {
                 while (resultSet.next()) {
-                    collectPrivileges(privilegeCache, resultSet);
+                    collectTablePrivileges(privilegeCache, resultSet);
                 }
             }
         }
         fillTablePrivileges(privilegeCache, userPrivilegeMap);
     }
-    
+
     private void fillTablePrivileges(final Map<ShardingSphereUser, Map<String, Map<String, List<PrivilegeType>>>> privilegeCache, 
                                      final Map<ShardingSphereUser, ShardingSpherePrivileges> userPrivilegeMap) {
         for (Entry<ShardingSphereUser, Map<String, Map<String, List<PrivilegeType>>>> entry : privilegeCache.entrySet()) {
@@ -89,13 +87,13 @@ public final class PostgreSQLPrivilegeLoader implements StoragePrivilegeLoader {
             }
         }
     }
-    
-    private void collectPrivileges(final Map<ShardingSphereUser, Map<String, Map<String, List<PrivilegeType>>>> privilegeCache, final ResultSet resultSet) throws SQLException {
-        String db = resultSet.getString("table_catalog");
-        String tableName = resultSet.getString("table_name");
-        String privilegeType = resultSet.getString("privilege_type");
-        boolean hasPrivilege = resultSet.getString("is_grantable").equalsIgnoreCase("TRUE");
-        String grantee = resultSet.getString("grantee");
+
+    private void collectTablePrivileges(final Map<ShardingSphereUser, Map<String, Map<String, List<PrivilegeType>>>> privilegeCache, final ResultSet resultSet) throws SQLException {
+        String db = resultSet.getString("TABLE_SCHEMA");
+        String tableName = resultSet.getString("TABLE_NAME");
+        String privilegeType = resultSet.getString("PRIVILEGE");
+        boolean hasPrivilege = "YES".equalsIgnoreCase(resultSet.getString("GRANTABLE"));
+        String grantee = resultSet.getString("GRANTEE");
         if (hasPrivilege) {
             privilegeCache
                     .computeIfAbsent(new ShardingSphereUser(grantee, "", ""), k -> new HashMap<>())
@@ -104,52 +102,50 @@ public final class PostgreSQLPrivilegeLoader implements StoragePrivilegeLoader {
                     .add(getPrivilegeType(privilegeType));
         }
     }
-    
-    private void fillRolePrivileges(final Map<ShardingSphereUser, ShardingSpherePrivileges> userPrivilegeMap, 
-                                    final DataSource dataSource, final Collection<ShardingSphereUser> users) throws SQLException {
+
+    private void fillSysPrivileges(final Map<ShardingSphereUser, ShardingSpherePrivileges> userPrivilegeMap, final DataSource dataSource, 
+                                   final Collection<ShardingSphereUser> users) throws SQLException {
+        Map<ShardingSphereUser, List<PrivilegeType>> privilegeCache = new HashMap<>();
         try (Connection connection = dataSource.getConnection()) {
             Statement statement = connection.createStatement();
-            try (ResultSet resultSet = statement.executeQuery(getRolePrivilegesSQL(users))) {
+            try (ResultSet resultSet = statement.executeQuery(getSysPrivilegesSQL(users))) {
                 while (resultSet.next()) {
-                    fillRolePrivileges(userPrivilegeMap, resultSet);
+                    collectSysPrivileges(privilegeCache, resultSet);
                 }
             }
         }
+        fillSysPrivileges(privilegeCache, userPrivilegeMap);
     }
-    
-    private void fillRolePrivileges(final Map<ShardingSphereUser, ShardingSpherePrivileges> userPrivilegeMap, final ResultSet resultSet) throws SQLException {
-        Optional<ShardingSphereUser> user = findShardingSphereUser(userPrivilegeMap, resultSet);
-        if (user.isPresent()) {
-            userPrivilegeMap.get(user.get()).getAdministrativePrivileges().getPrivileges().addAll(loadRolePrivileges(resultSet));
+
+    private void fillSysPrivileges(final Map<ShardingSphereUser, List<PrivilegeType>> privilegeCache, final Map<ShardingSphereUser, ShardingSpherePrivileges> userPrivilegeMap) throws SQLException {
+        for (Entry<ShardingSphereUser, List<PrivilegeType>> entry : privilegeCache.entrySet()) {
+            userPrivilegeMap.get(entry.getKey()).getAdministrativePrivileges().getPrivileges().addAll(entry.getValue());
         }
     }
-    
+
+    private void collectSysPrivileges(final Map<ShardingSphereUser, List<PrivilegeType>> privilegeCache, final ResultSet resultSet) throws SQLException {
+        String privilegeType = resultSet.getString("PRIVILEGE");
+        String grantee = resultSet.getString("GRANTEE");
+        privilegeCache
+                .computeIfAbsent(new ShardingSphereUser(grantee, "", ""), k -> new ArrayList<>())
+                .add(getPrivilegeType(privilegeType));
+    }
+
     private Optional<ShardingSphereUser> findShardingSphereUser(final Map<ShardingSphereUser, ShardingSpherePrivileges> userPrivilegeMap, final ResultSet resultSet) throws SQLException {
         Grantee grantee = new Grantee(resultSet.getString("rolname"), "");
         return userPrivilegeMap.keySet().stream().filter(each -> each.getGrantee().equals(grantee)).findFirst();
     }
-    
-    private Collection<PrivilegeType> loadRolePrivileges(final ResultSet resultSet) throws SQLException {
-        Collection<PrivilegeType> result = new LinkedList<>();
-        addToPrivilegeTypesIfPresent(resultSet.getBoolean("rolsuper"), PrivilegeType.SUPER, result);
-        addToPrivilegeTypesIfPresent(resultSet.getBoolean("rolcreaterole"), PrivilegeType.CREATE_ROLE, result);
-        addToPrivilegeTypesIfPresent(resultSet.getBoolean("rolcreatedb"), PrivilegeType.CREATE_DATABASE, result);
-        addToPrivilegeTypesIfPresent(resultSet.getBoolean("rolreplication"), PrivilegeType.REPL_CLIENT, result);
-        addToPrivilegeTypesIfPresent(resultSet.getBoolean("rolinherit"), PrivilegeType.INHERIT, result);
-        addToPrivilegeTypesIfPresent(resultSet.getBoolean("rolcanlogin"), PrivilegeType.CAN_LOGIN, result);
-        return result;
+
+    private String getSysPrivilegesSQL(final Collection<ShardingSphereUser> users) {
+        String userList = users.stream().map(each -> String.format("'%s'", each.getGrantee().getUsername())).collect(Collectors.joining(", "));
+        return String.format(SYS_PRIVILEGE_SQL, userList);
     }
-    
+
     private String getTablePrivilegesSQL(final Collection<ShardingSphereUser> users) {
         String userList = users.stream().map(each -> String.format("'%s'", each.getGrantee().getUsername())).collect(Collectors.joining(", "));
         return String.format(TABLE_PRIVILEGE_SQL, userList);
     }
-    
-    private String getRolePrivilegesSQL(final Collection<ShardingSphereUser> users) {
-        String userList = users.stream().map(each -> String.format("'%s'", each.getGrantee().getUsername())).collect(Collectors.joining(", "));
-        return String.format(ROLES_SQL, userList);
-    }
-    
+
     private PrivilegeType getPrivilegeType(final String privilege) {
         switch (privilege) {
             case "SELECT":
@@ -160,35 +156,34 @@ public final class PostgreSQLPrivilegeLoader implements StoragePrivilegeLoader {
                 return PrivilegeType.UPDATE;
             case "DELETE":
                 return PrivilegeType.DELETE;
-            case "TRUNCATE":
-                return PrivilegeType.TRUNCATE;
             case "REFERENCES":
                 return PrivilegeType.REFERENCES;
-            case "TRIGGER":
-                return PrivilegeType.TRIGGER;
-            case "CREATE":
-                return PrivilegeType.CREATE;
+            case "INDEX":
+                return PrivilegeType.INDEX;
             case "EXECUTE":
                 return PrivilegeType.EXECUTE;
-            case "USAGE":
-                return PrivilegeType.USAGE;
-            case "CONNECT":
-                return PrivilegeType.CONNECT;
-            case "TEMPORARY":
-                return PrivilegeType.TEMPORARY;
+            case "CREATE PROCEDURE":
+                return PrivilegeType.CREATE_PROC;
+            case "CREATE ROLE":
+                return PrivilegeType.CREATE_ROLE;
+            case "CREATE SEQUENCE":
+                return PrivilegeType.CREATE_SEQUENCE;
+            case "CREATE TABLESPACE":
+                return PrivilegeType.CREATE_TABLESPACE;
+            case "CREATE USER":
+                return PrivilegeType.CREATE_USER;
+            case "CREATE VIEW":
+                return PrivilegeType.CREATE_VIEW;
+            case "SYSDBA":
+                return PrivilegeType.SUPER;
+            // TODO other privilege
             default:
                 throw new UnsupportedOperationException(privilege);
         }
     }
-    
-    private void addToPrivilegeTypesIfPresent(final boolean hasPrivilege, final PrivilegeType privilegeType, final Collection<PrivilegeType> target) {
-        if (hasPrivilege) {
-            target.add(privilegeType);
-        }
-    }
-    
+
     @Override
     public String getType() {
-        return "PostgreSQL";
+        return "Oracle";
     }
 }
