+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-common/src/main/resources/META-INF/services/org.apache.shardingsphere.authority.algorithm.natived.loader.StoragePrivilegeLoader
@@ -15,6 +15,6 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.authority.algorithm.storage.loader.impl.dialect.MySQLPrivilegeLoader
-org.apache.shardingsphere.authority.algorithm.storage.loader.impl.dialect.PostgreSQLPrivilegeLoader
-org.apache.shardingsphere.authority.algorithm.storage.loader.impl.dialect.OraclePrivilegeLoader
+org.apache.shardingsphere.authority.algorithm.natived.loader.impl.dialect.MySQLPrivilegeLoader
+org.apache.shardingsphere.authority.algorithm.natived.loader.impl.dialect.PostgreSQLPrivilegeLoader
+org.apache.shardingsphere.authority.algorithm.natived.loader.impl.dialect.OraclePrivilegeLoader
