+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-common/src/main/resources/META-INF/services/org.apache.shardingsphere.authority.spi.AuthorityProvideAlgorithm
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.authority.algorithm.storage.StorageAuthorityProviderAlgorithm
+org.apache.shardingsphere.authority.algorithm.natived.NativeAuthorityProviderAlgorithm
