+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-common/src/test/java/org/apache/shardingsphere/authority/algorithm/natived/loader/impl/dialect/PostgreSQLPrivilegeLoaderTest.java
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.authority.algorithm.storage.loader.impl.dialect;
+package org.apache.shardingsphere.authority.algorithm.natived.loader.impl.dialect;
 
-import org.apache.shardingsphere.authority.algorithm.storage.loader.StoragePrivilegeLoader;
 import org.apache.shardingsphere.authority.model.PrivilegeType;
 import org.apache.shardingsphere.authority.model.ShardingSpherePrivileges;
 import org.apache.shardingsphere.authority.model.database.SchemaPrivileges;
+import org.apache.shardingsphere.authority.algorithm.natived.loader.StoragePrivilegeLoader;
 import org.apache.shardingsphere.infra.metadata.user.ShardingSphereUser;
 import org.apache.shardingsphere.infra.spi.ShardingSphereServiceLoader;
 import org.apache.shardingsphere.infra.spi.typed.TypedSPIRegistry;
@@ -48,7 +48,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public final class OraclePrivilegeLoaderTest {
+public final class PostgreSQLPrivilegeLoaderTest {
     
     @BeforeClass
     public static void setUp() {
@@ -59,57 +59,63 @@ public final class OraclePrivilegeLoaderTest {
     public void assertLoad() throws SQLException {
         Collection<ShardingSphereUser> users = createUsers();
         DataSource dataSource = mockDataSource(users);
-        assertPrivileges(TypedSPIRegistry.getRegisteredService(StoragePrivilegeLoader.class, "Oracle", new Properties()).load(users, dataSource));
+        assertPrivileges(TypedSPIRegistry.getRegisteredService(StoragePrivilegeLoader.class, "PostgreSQL", new Properties()).load(users, dataSource));
     }
     
     private void assertPrivileges(final Map<ShardingSphereUser, ShardingSpherePrivileges> actual) {
         assertThat(actual.size(), is(1));
-        ShardingSphereUser user = new ShardingSphereUser("admin", "", "");
+        ShardingSphereUser user = new ShardingSphereUser("postgres", "", "");
         assertThat(actual.get(user).getDatabasePrivileges().getGlobalPrivileges().size(), is(0));
         assertThat(actual.get(user).getDatabasePrivileges().getSpecificPrivileges().size(), is(1));
-        Collection<PrivilegeType> expectedSpecificPrivilege = new CopyOnWriteArraySet(Arrays.asList(PrivilegeType.INSERT, PrivilegeType.SELECT, PrivilegeType.UPDATE));
-        SchemaPrivileges schemaPrivileges = actual.get(user).getDatabasePrivileges().getSpecificPrivileges().get("sys");
+        Collection<PrivilegeType> expectedSpecificPrivilege = new CopyOnWriteArraySet(Arrays.asList(PrivilegeType.INSERT, PrivilegeType.SELECT, PrivilegeType.UPDATE,
+                PrivilegeType.DELETE));
+        SchemaPrivileges schemaPrivileges = actual.get(user).getDatabasePrivileges().getSpecificPrivileges().get("db0");
         assertThat(schemaPrivileges.getSpecificPrivileges().get("t_order").hasPrivileges(expectedSpecificPrivilege), is(true));
-        assertThat(actual.get(user).getAdministrativePrivileges().getPrivileges().size(), is(3));
+        assertThat(actual.get(user).getAdministrativePrivileges().getPrivileges().size(), is(4));
         Collection<PrivilegeType> expectedAdministrativePrivilege = new CopyOnWriteArraySet(Arrays.asList(PrivilegeType.SUPER, PrivilegeType.CREATE_ROLE,
-                PrivilegeType.CREATE_TABLESPACE));
+                PrivilegeType.CREATE_DATABASE, PrivilegeType.CAN_LOGIN));
         assertEquals(actual.get(user).getAdministrativePrivileges().getPrivileges(), expectedAdministrativePrivilege);
     }
-
+    
     private Collection<ShardingSphereUser> createUsers() {
         LinkedList<ShardingSphereUser> result = new LinkedList<>();
-        result.add(new ShardingSphereUser("admin", "", ""));
+        result.add(new ShardingSphereUser("postgres", "", ""));
         return result;
     }
-
+    
     private DataSource mockDataSource(final Collection<ShardingSphereUser> users) throws SQLException {
-        ResultSet sysPrivilegeResultSet = mockSysPrivilegeResultSet();
+        ResultSet tablePrivilegeResultSet = mockTablePrivilegeResultSet();
         DataSource result = mock(DataSource.class, RETURNS_DEEP_STUBS);
-        String sysPrivilegeSql = "SELECT GRANTEE, PRIVILEGE, ADMIN_OPTION, INHERITED FROM DBA_SYS_PRIVS WHERE GRANTEE IN (%s)";
+        String tablePrivilegeSql = "SELECT grantor, grantee, table_catalog, table_name, privilege_type, is_grantable from information_schema.table_privileges WHERE grantee IN (%s)";
         String userList = users.stream().map(item -> String.format("'%s'", item.getGrantee().getUsername(), item.getGrantee().getHostname())).collect(Collectors.joining(", "));
-        when(result.getConnection().createStatement().executeQuery(String.format(sysPrivilegeSql, userList))).thenReturn(sysPrivilegeResultSet);
-        ResultSet tabPrivilegeResultSet = mockTabPrivilegeResultSet();
-        String tabPrivilegeSql = "SELECT GRANTEE, TABLE_SCHEMA, TABLE_NAME, PRIVILEGE, GRANTABLE, INHERITED FROM ALL_TAB_PRIVS WHERE GRANTEE IN (%s)";
-        when(result.getConnection().createStatement().executeQuery(String.format(tabPrivilegeSql, userList))).thenReturn(tabPrivilegeResultSet);
+        when(result.getConnection().createStatement().executeQuery(String.format(tablePrivilegeSql, userList))).thenReturn(tablePrivilegeResultSet);
+        ResultSet rolePrivilegeResultSet = mockRolePrivilegeResultSet();
+        String rolePrivilegeSql = "select * from pg_roles WHERE rolname IN (%s)";
+        when(result.getConnection().createStatement().executeQuery(String.format(rolePrivilegeSql, userList))).thenReturn(rolePrivilegeResultSet);
         return result;
     }
-
-    private ResultSet mockSysPrivilegeResultSet() throws SQLException {
+    
+    private ResultSet mockTablePrivilegeResultSet() throws SQLException {
         ResultSet result = mock(ResultSet.class, RETURNS_DEEP_STUBS);
-        when(result.next()).thenReturn(true, true, true, false);
-        when(result.getString("GRANTEE")).thenReturn("admin");
-        when(result.getString("PRIVILEGE")).thenReturn("SYSDBA", "CREATE ROLE", "CREATE TABLESPACE");
+        when(result.next()).thenReturn(true, true, true, true, true, true, true, false);
+        when(result.getString("table_catalog")).thenReturn("db0");
+        when(result.getString("table_name")).thenReturn("t_order");
+        when(result.getString("privilege_type")).thenReturn("INSERT", "SELECT", "UPDATE", "DELETE", "TRUNCATE", "REFERENCES", "TRIGGER");
+        when(result.getString("is_grantable")).thenReturn("TRUE", "TRUE", "TRUE", "TRUE", "TRUE", "TRUE", "TRUE");
+        when(result.getString("grantee")).thenReturn("postgres");
         return result;
     }
-
-    private ResultSet mockTabPrivilegeResultSet() throws SQLException {
+    
+    private ResultSet mockRolePrivilegeResultSet() throws SQLException {
         ResultSet result = mock(ResultSet.class, RETURNS_DEEP_STUBS);
-        when(result.next()).thenReturn(true, true, true, true, false);
-        when(result.getString("TABLE_SCHEMA")).thenReturn("sys");
-        when(result.getString("TABLE_NAME")).thenReturn("t_order");
-        when(result.getString("PRIVILEGE")).thenReturn("SELECT", "INSERT", "DELETE", "UPDATE");
-        when(result.getString("GRANTABLE")).thenReturn("YES", "YES", "FALSE", "YES");
-        when(result.getString("GRANTEE")).thenReturn("admin");
+        when(result.next()).thenReturn(true, false);
+        when(result.getString("rolname")).thenReturn("postgres");
+        when(result.getBoolean("rolsuper")).thenReturn(true);
+        when(result.getBoolean("rolcreaterole")).thenReturn(true);
+        when(result.getBoolean("rolcreatedb")).thenReturn(true);
+        when(result.getBoolean("rolreplication")).thenReturn(false);
+        when(result.getBoolean("rolinherit")).thenReturn(false);
+        when(result.getBoolean("rolcanlogin")).thenReturn(true);
         return result;
     }
 }
