+++ shardingsphere-governance/shardingsphere-governance-context/src/main/java/org/apache/shardingsphere/governance/context/authority/GovernanceAuthorityContext.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.governance.context.authority;
 import com.google.common.eventbus.Subscribe;
 import lombok.Setter;
 import org.apache.shardingsphere.authority.AuthorityContext;
-import org.apache.shardingsphere.authority.algorithm.storage.StorageAuthorityProviderAlgorithm;
+import org.apache.shardingsphere.authority.algorithm.natived.NativeAuthorityProviderAlgorithm;
 import org.apache.shardingsphere.authority.spi.AuthorityProvideAlgorithm;
 import org.apache.shardingsphere.governance.core.event.model.auth.PrivilegeChangedEvent;
 import org.apache.shardingsphere.governance.core.event.model.auth.UserRuleChangedEvent;
@@ -61,7 +61,7 @@ public final class GovernanceAuthorityContext implements MetaDataAwareEventSubsc
 
     private void reloadAuthority(final Collection<ShardingSphereUser> users) {
         // TODO reload AuthorityCheckAlgorithm from SPI
-        AuthorityProvideAlgorithm result = new StorageAuthorityProviderAlgorithm();
+        AuthorityProvideAlgorithm result = new NativeAuthorityProviderAlgorithm();
         result.init(metaDataContexts.getMetaDataMap(), users);
         AuthorityContext.getInstance().init(result);
     }
