+++ shardingsphere-proxy/shardingsphere-proxy-bootstrap/src/main/resources/conf/server.yaml
@@ -60,4 +60,4 @@
 #rules:
 #  - !AUTHORITY
 #    provider:
-#      type: STORAGE
+#      type: NATIVE
